new rank_vf.py file, vector_calculus refactors
